+++ docs/CHANGELOG.rst
@@ -19,6 +19,7 @@ Changed
 - **BACKWARD INCOMPATIBLE:** Require Resolwe 37.x
 - Add ``sample_annotation`` Jinja expressions filter that accepts an annotation
   path and returns its value
+- Update field paths for sample annotation in ``geo-import`` process
 
 Fixed
 -----